added friends searching controls
