+++ mobile/app/search.tsx
@@ -1,0 +1,1 @@
+export { default } from '@components/screens/SCRSearch';
