+++ mobile/components/containers/index.ts
@@ -1,3 +1,3 @@
-export { default as CNTHomeScreenFeed } from './CNTHomeScreenFeed';
-export { default as CNTPostCard } from './CNTPostCard';
-export { default as CNTPostCardPlaceholder } from './CNTPostCardPlaceholder';
+export { default as CNTHomeScreenFeed } from "./CNTHomeScreenFeed";
+export { default as CNTPostCard } from "./CNTPostCard";
+export { default as CNTPostCardPlaceholder } from "./CNTPostCardPlaceholder";
