+++ mobile/components/controls/index.ts
@@ -1,6 +1,6 @@
-export { default as CTRLActionButton } from './CTRLActionButton';
-export { default as CTRLFormInput } from './CTRLFormInput';
-export { default as CTRLFormToggle } from './CTRLFormToggle';
-export { default as CTRLIcon } from './CTRLIcon';
-export { default as CTRLImageButton } from './CTRLImageButton';
-export { default as CTRLTabButton } from './CTRLTabButton';
+export { default as CTRLActionButton } from "./CTRLActionButton";
+export { default as CTRLFormInput } from "./CTRLFormInput";
+export { default as CTRLFormToggle } from "./CTRLFormToggle";
+export { default as CTRLIcon } from "./CTRLIcon";
+export { default as CTRLImageButton } from "./CTRLImageButton";
+export { default as CTRLTabButton } from "./CTRLTabButton";
