+++ mobile/components/screens/index.ts
@@ -1,0 +1,2 @@
+export { default as SCRHome } from './SCRHome';
+export { default as SCRSearch } from './SCRSearch';
