+++ mobile/context/index.ts
@@ -1,7 +1,7 @@
-export * from './AppProvider';
-export * from './AuthProvider';
-export * from './DataProvider';
-export * from './EnvironmentProvider';
-export * from './RootProvider';
-export * from './ThemeProvider';
-export * from './TranslationProvider';
+export * from "./AppProvider";
+export * from "./AuthProvider";
+export * from "./DataProvider";
+export * from "./EnvironmentProvider";
+export * from "./RootProvider";
+export * from "./ThemeProvider";
+export * from "./TranslationProvider";
