+++ mobile/navigation/NVFriendStackLayout.tsx
@@ -1,11 +1,17 @@
 import { Stack } from 'expo-router';
 
+import { SCRFriendsHomeHeader } from '@components/screens/SCRFriendsHome';
+
 export default function NVFriendStackLayout(): JSX.Element {
   return (
-    <Stack
-      screenOptions={{
-        headerShown: false,
-      }}
-    />
+    <Stack>
+      <Stack.Screen
+        name="index"
+        options={{
+          headerShown: true,
+          header: () => <SCRFriendsHomeHeader />,
+        }}
+      />
+    </Stack>
   );
 }
